saas-tenant-isolation-abac-jwt-sts/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
auth0-python==4.4.0
    # via -r requirements.in
aws-cdk-lib==2.90.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-lambda-python-alpha
    #   cdk-nag
aws-cdk.asset-awscli-v1==2.2.200
    # via aws-cdk-lib
aws-cdk.asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk.asset-node-proxy-agent-v5==2.0.166
    # via aws-cdk-lib
aws-cdk.aws-lambda-python-alpha==2.90.0a0
    # via -r requirements.in
cattrs==23.1.2
    # via jsii
cdk-nag==2.27.95
    # via -r requirements.in
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
constructs==10.2.69
    # via
    #   -r requirements.in
    #   aws-cdk-lib
    #   aws-cdk.aws-lambda-python-alpha
    #   cdk-nag
cryptography==41.0.3
    # via pyjwt
exceptiongroup==1.1.2
    # via cattrs
idna==3.4
    # via requests
importlib-resources==6.0.1
    # via jsii
jsii==1.86.1
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v5
    #   aws-cdk.aws-lambda-python-alpha
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v5
    #   aws-cdk.aws-lambda-python-alpha
    #   cdk-nag
    #   constructs
    #   jsii
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via auth0-python
python-dateutil==2.8.2
    # via jsii
requests==2.31.0
    # via auth0-python
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-lib
    #   aws-cdk.asset-awscli-v1
    #   aws-cdk.asset-kubectl-v20
    #   aws-cdk.asset-node-proxy-agent-v5
    #   aws-cdk.aws-lambda-python-alpha
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.7.1
    # via
    #   cattrs
    #   jsii
urllib3==2.0.4
    # via requests
zipp==3.16.2
    # via importlib-resources
